Update and fix bitmap

+++ base/rgb.h
@@ -96,9 +96,12 @@ H_INLINE int MMRGBHexSimilarToColor(MMRGBHex h1, MMRGBHex h2, float tolerance)
 	if (tolerance <= 0.0f) {
 		return h1 == h2;
 	} else {
-		uint8_t d1 = RED_FROM_HEX(h1) - RED_FROM_HEX(h2);
-		uint8_t d2 = GREEN_FROM_HEX(h1) - GREEN_FROM_HEX(h2);
-		uint8_t d3 = BLUE_FROM_HEX(h1) - BLUE_FROM_HEX(h2);
+		// uint8_t d1 = RED_FROM_HEX(h1) - RED_FROM_HEX(h2);
+		// uint8_t d2 = GREEN_FROM_HEX(h1) - GREEN_FROM_HEX(h2);
+		// uint8_t d3 = BLUE_FROM_HEX(h1) - BLUE_FROM_HEX(h2);
+		int d1 = RED_FROM_HEX(h1) - RED_FROM_HEX(h2);
+		int d2 = GREEN_FROM_HEX(h1) - GREEN_FROM_HEX(h2);
+		int d3 = BLUE_FROM_HEX(h1) - BLUE_FROM_HEX(h2);
 		return sqrt((double)(d1 * d1) +
 		            (d2 * d2) +
 		            (d3 * d3)) <= (tolerance * 442.0f);
